SPORTS-193 Rename file
