+++ api/src/mutations/index.ts
@@ -2,7 +2,7 @@ import { login } from './login';
 import { signUp } from './signUp';
 import { createGame, editGame, deleteGame } from './game';
 import { refreshAccessToken } from './refreshAccessToken';
-import { createRound, editRound, deleteRound } from './rounds';
+import { createRound, editRound, deleteRound } from './round';
 import { createCategory, editCategory, deleteCategory } from './category';
 
 export default {
